fraud_detector/requirements.txt
requests
scikit-learn
solace-pubsubplus
python-dotenv